feat: put and delete now recognises if a id exist

+++ routes/books.js
@@ -32,7 +32,7 @@ router.post('/', (req, res) => {
 
   const newBook = { Titel, Erscheinungsjahr, Autor };
 
-  connection.query('INSERT INTO Books SET ?', newBook, (err, result) => {
+  connection.query('INSERT INTO Books SET ?', newBook, (err) => {
     if (err) throw err;
     res.status(201).send('Buch erfolgreich erstellt');
   });
@@ -46,24 +46,42 @@ router.put('/:id', (req, res) => {
     return res.status(400).send("Titel, Erscheinungsjahr und Autor sind erforderlich");
   }
 
-  const query = `UPDATE Books SET Titel = ?, Erscheinungsjahr = ?, Autor = ? WHERE Id = ?`;
-
-  connection.query(query, [Titel, Erscheinungsjahr, Autor, bookId], (err) => {
+  connection.query('SELECT * FROM Books WHERE Id = ?', [bookId], (err, results) => {
     if (err) throw err;
-    res.send("Buch erfolgreich aktualisiert");
+
+    if (results.length > 0) {
+      const query = `UPDATE Books SET Titel = ?, Erscheinungsjahr = ?, Autor = ? WHERE Id = ?`;
+
+      connection.query(query, [Titel, Erscheinungsjahr, Autor, bookId], (err) => {
+        if (err) throw err;
+        res.send("Buch erfolgreich aktualisiert");
+      });
+    } else {
+      res.status(404).send("Buch nicht gefunden");
+    }
   });
 });
+
 
 router.delete('/:id', (req, res) => {
   const bookId = req.params.id;
 
-  const query = `DELETE FROM Books WHERE Id = ?`;
-
-  connection.query(query, [bookId], (err) => {
+  connection.query('SELECT * FROM Books WHERE Id = ?', [bookId], (err, results) => {
     if (err) throw err;
-    res.send("Buch erfolgreich gelöscht");
+
+    if (results.length > 0) {
+      const query = `DELETE FROM Books WHERE Id = ?`;
+
+      connection.query(query, [bookId], (err) => {
+        if (err) throw err;
+        res.send("Buch erfolgreich gelöscht");
+      });
+    } else {
+      res.status(404).send("Buch nicht gefunden"); 
+    }
   });
 });
+
